+++ src/main/java/net/runelite/client/plugins/microbot/miningplus/AutoMiningPlusPlugin.java
@@ -25,7 +25,7 @@ import java.awt.*;
 )
 @Slf4j
 public class AutoMiningPlusPlugin extends Plugin {
-    public static final String version = "0.5.11";
+    public static final String version = "0.5.12";
     @Inject
     private AutoMiningPlusConfig config;
     @Provides